backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pandas==2.2.3
numpy==2.1.3
google-generativeai==0.3.0
python-multipart==0.0.6